requirements.txt
python-twitter
numpy
matplotlib

